Fix compiler warning
src/winscard_msg.c:249:17: warning: implicit conversion loses integer precision: 'ssize_t' (aka 'long') to 'int' [-Wshorten-64-to-32]
                        bytes_read = read(filedes, buffer, remaining);
                                   ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ src/winscard_msg.c
@@ -238,7 +238,7 @@ INTERNAL LONG MessageReceiveTimeout(uint32_t command, void *buffer_void,
 		/* try to read only when socket is readable */
 		if (pollret > 0)
 		{
-			int bytes_read;
+			ssize_t bytes_read;
 
 			if (!(read_fd.revents & POLLIN))
 			{
